backend/requirements.txt
Flask
flask-cors
google-generativeai
deep-translator
python-dotenv
gunicorn
